Model refactoring
