+++ src/app/model/poll-item-answer-participant/multiple-choice-item-answer-participant.ts
