+++ src/app/model/poll-item-answer-participant/open-text-item-answer-participant.ts
