+++ src/app/model/poll-item-answer-participant/quiz-item-answer-participant.ts
