+++ src/app/model/poll-item-answer/multiple-choice-item-answer.ts
