+++ src/app/model/poll-item-answer/open-text-item-answer.ts
