+++ src/app/model/poll-item-answer/quiz-item-answer.ts
