+++ src/app/model/poll-item/open-text-item.ts
