+++ src/app/model/poll-item/poll-item.ts
