+++ src/app/pages/presentation/presentation.component.ts
@@ -4,9 +4,9 @@
 
 import {Component, OnInit} from '@angular/core';
 import {Poll} from '../../model/poll';
-import {MultipleChoiceItem} from '../../model/multiple-choice-item';
-import {QuizItem} from '../../model/quiz-item';
-import {OpenTextItem} from '../../model/open-text-item';
+import {MultipleChoiceItem} from '../../model/poll-item/multiple-choice-item';
+import {QuizItem} from '../../model/poll-item/quiz-item';
+import {OpenTextItem} from '../../model/poll-item/open-text-item';
 import {ActivatedRoute, Router} from '@angular/router';
 import {PollService} from '../../service/poll.service';
 import {WebsocketService} from '../../service/websocket.service';
